Split impl Parse for v1::Linear to a separate file
